Building blocks
Created the classes "ChanConfig" (stores channel configurations) and "Measurement" (stores and processes the data from one file).

Also created the main file, which takes and filters user input to create charts according to needs and user specifications.

In this version, all charts are displayed on top of each other on one graph.